Perf for the win

+++ packages/rules/src/fileContents.ts
@@ -15,21 +15,21 @@ const Options = r.Union(
   r.Record({
     file: r.String,
     generator: r.Function,
-    template: r.Undefined,
-    templateFile: r.Undefined,
+    template: r.Undefined.optional(),
+    templateFile: r.Undefined.optional(),
   }),
 
   r.Record({
     file: r.String,
-    generator: r.Undefined,
+    generator: r.Undefined.optional(),
     template: r.String,
-    templateFile: r.Undefined,
+    templateFile: r.Undefined.optional(),
   }),
 
   r.Record({
     file: r.String,
-    generator: r.Undefined,
-    template: r.Undefined,
+    generator: r.Undefined.optional(),
+    template: r.Undefined.optional(),
     templateFile: r.String,
   })
 );
@@ -39,8 +39,7 @@ type Options = r.Static<typeof Options>;
 export const fileContents = {
   check: function expectFileContents(context: Context, opts: Options) {
     const fullPath = path.join(context.packageDir, opts.file);
-    const generator = getGenerator(context, opts);
-    const expectedContent = generator(context);
+    const expectedContent = getExpectedContents(context, opts);
 
     const pathExists = context.host.exists(fullPath);
     const actualContent = pathExists ? context.host.readFile(fullPath, { encoding: "utf-8" }) : undefined;
@@ -50,11 +49,11 @@ export const fileContents = {
         message: "Expect file contents to match",
         longMessage: diff(expectedContent, actualContent, { expand: true }),
         fixer: () => {
-          if (expectedContent === undefined && pathExists) {
-            context.host.deleteFile(fullPath);
+          if (expectedContent === undefined) {
+            if (pathExists) context.host.deleteFile(fullPath);
           } else {
             context.host.mkdir(path.dirname(fullPath), { recursive: true });
-            context.host.writeFile(fullPath, expectedContent);
+            context.host.writeFile(fullPath, expectedContent, { encoding: "utf-8" });
           }
         },
       });
@@ -63,25 +62,30 @@ export const fileContents = {
   optionsRuntype: Options,
 } as RuleModule<typeof Options>;
 
-function getGenerator(context: Context, opts: Options) {
+const optionsCache = new Map<Options, ((context: Context) => string | undefined) | string | undefined>();
+
+function getExpectedContents(context: Context, opts: Options) {
+  // we need to use has because undefined is a valid value in the cache
+  if (optionsCache.has(opts)) {
+    const cachedEntry = optionsCache.get(opts);
+    if (cachedEntry && typeof cachedEntry === "function") {
+      return cachedEntry(context);
+    }
+    return cachedEntry;
+  }
+
   if (opts.generator) {
-    return opts.generator;
+    optionsCache.set(opts, opts.generator);
+    return opts.generator(context) as string | undefined; // we have no guarentee its the right kind of function
   } else if (opts.templateFile) {
     const { packageDir: workspacePackageDir } = context.getWorkspaceContext();
     const fullPath = path.resolve(workspacePackageDir, opts.templateFile);
     const template = context.host.readFile(fullPath, { encoding: "utf-8" });
 
-    return makeGenerator(template);
-  } else if (opts.template) {
-    return makeGenerator(opts.template);
-  } else {
-    throw new Error("Unable to make generator");
-  }
-}
-
-function makeGenerator(template: string) {
-  // tslint:disable-next-line:variable-name
-  return function generator(_context: Context) {
+    optionsCache.set(opts, template);
     return template;
-  };
+  } else {
+    optionsCache.set(opts, opts.template);
+    return opts.template;
+  }
 }
